plugins: install: factor out plugin archive fetching
